requirements.txt
requests==2.31.0
openai==1.3.0
mido==1.3.0
numpy==1.24.3
streamlit==1.28.0
python-dotenv==1.0.0
pydub==0.25.1
